tmcprototype/sdpmasterleafnode/requirements.txt
numpy == 1.17.2
pytango >= 9.3.2
jsonschema >= 3.2.0
ska-logging==0.3.0
lmcbaseclasses == 0.7.2
marshmallow
cdm-shared-library
ska-telescope-model
fire
skatmccommon==0.1.1+4fba2733